backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database & ORM
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Data Processing
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
xlrd==2.0.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.0
tabula-py==2.9.0

# AI & ML
openai==1.14.3
sentence-transformers==2.6.1
faiss-cpu==1.8.0
scikit-learn==1.4.2

# Azure Services
azure-ai-textanalytics==5.3.0
azure-storage-blob==12.19.0

# Cache & Message Queue
redis==5.0.3
celery==5.3.6

# File & Image Processing
aiofiles==23.2.0
Pillow==10.3.0

# Data Visualization
matplotlib==3.8.4
seaborn==0.13.2

# HTTP & API
httpx==0.27.0
websockets==12.0

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1

# Utilities
email-validator==2.1.1
python-dateutil==2.9.0.post0
jinja2==3.1.3

# Logging & Error Handling
structlog==24.1.0